SA44/124 integration
Added global static struct device_traits to solve the issue of the
program needing to know about the device from basically anywhere in the
program. Removing magic numbers and hard coded values from the BB60
stuff.

Last clean build for BB60 users is the official 2.0.5 commit.

+++ BBApp/src/lib/bb_lib.h
@@ -23,6 +23,8 @@
 #include "time_type.h"
 #include "kiss_fft/kissfft.hh"
 
+#include "lib/device_traits.h"
+
 class Trace;
 
 #define INDEX_OFFSET(x) ((GLvoid*)x)
@@ -300,7 +302,7 @@ void getPeakCorrelation(const complex_f *src,
                         double &peakPower);
 
 double getSignalFrequency(const std::vector<complex_f> &src, double sampleRate);
-QString getSampleRateString(int sampleRate);
+QString getSampleRateString(double sampleRate);
 
 namespace bb_lib {
